Refactor solution
