+++ public/app/plugins/datasource/loki/components/LokiQueryField.tsx
@@ -11,11 +11,11 @@ import {
 } from '@grafana/ui';
 import { Plugin, Node } from 'slate';
 import { LokiLabelBrowser } from './LokiLabelBrowser';
-import { ExploreQueryFieldProps, TimeRange } from '@grafana/data';
+import { ExploreQueryFieldProps } from '@grafana/data';
 import { LokiQuery, LokiOptions } from '../types';
 import { LanguageMap, languages as prismLanguages } from 'prismjs';
 import LokiLanguageProvider, { LokiHistoryItem } from '../language_provider';
-import { roundMsToMin } from '../language_utils';
+import { shouldRefreshLabels } from '../language_utils';
 import LokiDatasource from '../datasource';
 
 function getChooserText(hasSyntax: boolean, hasLogLabels: boolean) {
@@ -58,7 +58,6 @@ function willApplySuggestion(suggestion: string, { typeaheadContext, typeaheadTe
 export interface LokiQueryFieldProps extends ExploreQueryFieldProps<LokiDatasource, LokiQuery, LokiOptions> {
   history: LokiHistoryItem[];
   ExtraFieldElement?: ReactNode;
-  range?: TimeRange;
   placeholder?: string;
   'data-testid'?: string;
 }
@@ -98,22 +97,11 @@ export class LokiQueryField extends React.PureComponent<LokiQueryFieldProps, Lok
       range,
       datasource: { languageProvider },
     } = this.props;
-    const changedRangeToRefresh = this.rangeChangedToRefresh(range, prevProps.range);
+    const changedRangeToRefresh = shouldRefreshLabels(range, prevProps.range);
     // We want to refresh labels when range changes (we round up intervals to a minute)
     if (changedRangeToRefresh) {
-      console.log('here');
       await languageProvider.fetchLabels();
     }
-  }
-
-  rangeChangedToRefresh(range?: TimeRange, prevRange?: TimeRange): boolean {
-    if (range && prevRange) {
-      const sameMinuteFrom = roundMsToMin(range.from.valueOf()) === roundMsToMin(prevRange.from.valueOf());
-      const sameMinuteTo = roundMsToMin(range.to.valueOf()) === roundMsToMin(prevRange.to.valueOf());
-      // If both are same, don't need to refresh
-      return !(sameMinuteFrom && sameMinuteTo);
-    }
-    return false;
   }
 
   onChangeLabelBrowser = (selector: string) => {
