+++ public/app/plugins/datasource/loki/language_utils.ts
@@ -1,7 +1,19 @@
-export function roundMsToMin(milliseconds: number): number {
+import { TimeRange } from '@grafana/data';
+
+function roundMsToMin(milliseconds: number): number {
   return roundSecToMin(milliseconds / 1000);
 }
 
-export function roundSecToMin(seconds: number): number {
+function roundSecToMin(seconds: number): number {
   return Math.floor(seconds / 60);
+}
+
+export function shouldRefreshLabels(range?: TimeRange, prevRange?: TimeRange): boolean {
+  if (range && prevRange) {
+    const sameMinuteFrom = roundMsToMin(range.from.valueOf()) === roundMsToMin(prevRange.from.valueOf());
+    const sameMinuteTo = roundMsToMin(range.to.valueOf()) === roundMsToMin(prevRange.to.valueOf());
+    // If both are same, don't need to refresh
+    return !(sameMinuteFrom && sameMinuteTo);
+  }
+  return false;
 }
